requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydub>=0.25.1
torch>=2.0.0
python-multipart>=0.0.18
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
deepgram-sdk>=3.0.0
pyannote.audio>=3.1.0
numpy>=1.24.0
